Move requisition navigation out of hook
